Created Vivado project.
There's a great deal going on here but I'll try and summarise!

In Vivado 2018.3 I did the following:
* Create new project
* Set the project location by navigating to the close of this project
* In the file navigator create a new folder, 'Work_Dir', name taken from UG892
* Specify project same as repo
* Clear 'Create project subdirectory' as I've done it
* Project type is RTL
* On Add Sources page add 'Testbenches' directory and select 'Simulation only'
* Clear 'Copy sources into project' as I think this is what UG892 says when using CVS
* No constraints for now as not needed for simulation
* Default part is the KCU105 board I've got

After clicking Finish Vivado creates simple-fpga-cvs.xpr and various subdirectories. I belive, from UG892, that the xpr, and sources, is all that is needed to recreate the project. Hence ignore the /working/ directories.