Added IO port constraints after synthesising design.
The simulation was running which I took to mean the design is fine.

So moved onto synthesis by clicking Run Synthesis in the Flow Navigator.

A license is required for synthesis so I had to VPN to the office.

When synthesis had completed I clicked Schematic which showed a /schematic/ of my design which is just a buffer.

I clicked the '2 I/O Ports' link at the top of the schematic. This took me to a list of I/O ports in another tab.

In this table was a 'Package Pin' column. Referring to KCU105 Board User Guide UG917 Table 1-21 I selected the pins for LED0 and SW central.

I pressed CTRL-s and I was asked to save a constraints file. I used the file navigator to create ip-ports.xdc.

The 'simple-fpga-cvs.runs' directory appeared, I don't know what this means so I'm ignoring it.