Specify the I/O standard of I/O ports.
I clicked Run Implementation in the Flow Navigator and it appeared to succeed.

I then clicked Generate Bitstream but this failed with error about the I/O Std of the ports being 'Default'. I changed the I/O Std to LVCMOS18 because that's what it says in UG917. Now I have a bitstream!