Added 300 MHz oscillator to the testbench.
The KCU105 includes a 300 MHz oscillator. While I'm not sure what I want to do with it I've got a feeling it will be useful.

Before I can use it in the design it needs to be added to the testbench.

Oscillators are special, the have 2 pins, P and N.