Wired up the KCU105's 300 MHz to the FPGA, connected it to an SMA connector.
It appears that Vivado does some magic that I don't really understand for the simple GPIO, in the schematic the are IBUF things!

AFAICT, the same magic doesn't work for the oscillator inputs and a buffer has to be added in the design.

On the menu Tools|Language Templates I found the template code for the IBUFDS, "Verilog|Device Primitive Instantiation|Kintex Ultrascale|I/O|INPUT BUFFER|Differential Input Buffer".

I can see the oscillator on the user SMA. I was expecting to see a digital signal, i.e. square wave, but it turns out it is the oscillator waveform. Not sure what this means yet!