Made the MMCM clock output 10 MHz.
At 10 MHz I can see a square wave on the user SMA connector with the scope. At 50 MHz the waveform wasn't very square.

It's not clear to me if this was a problem with my design, the scope or the FPGA pins.

For now, keep it simple and use a lower frequency clock.