Added the oscillator input to the constraints.
It seems clock constraints are important but I don't really understand why or what to do with them.

I ran the Constraints Wizard and it offered this clock constraint to me. I guess it understood that these were oscillator inputs or they were the input to MMCM block.

The design ran before and it still runs now but I guess this would be important for the tools if the design was bigger.