Replaced MMCM will PLL because MMCM is overkill.
I ran 'RTL Analysis' in the Flow Navigator, because I read about it UltraFast Design Methodology Guide for the Vivado Design Suite UG949.

The 'Report Methodology' said I should use a PLL because it would be easier to place and an MMCM was unnecessary.

It don't think it reduced the number of cells in the synthesis schematic.

I had previously contemplated using a PLL so I didn't need much persuasion.