Define I/O bank 0 properties in constraints.
I ran the 'Report DRC' in the RTL Analysis Flow Navigator and got this warning:
    CFGBVS #1 Warning Neither the CFGBVS nor CONFIG_VOLTAGE voltage property is set in the current_design.

I looked in the KCU105 user manual and couldn't find any mention of the CFGBVS pin so I'm not certain how it is configured.

I did find an example project on the Internet that used these settings.

I re-generated the bitstream and re-programmed the device and it didn't blow up so let's hope I've done the correct thing.