Suppress warnings about unconnected design ports.
If this was C++ I'd be all over compiling without warnings.

The FPGA tools seem more /warning happy/ and I can't find ways to remove the warnings in code, like I'd do in C++.

I think [Synth 8-3331] is an innocuous warning, except the design might be suboptimal, so I've suppressed it in the project.