Suppress warning about not enough connections for library module.
I don't want to specify the PLL connections I'm not using and the design seems to work.

But, I also want to be warning free.

I couldn't find anything on the Internet to confirm that not giving all the connections is fine but I'm going to stick with it for now and suppress the warning in the project.