Made the PLL clock output 100 MHz.
I've been trying to debug my design using 'In-System Debugging' and couldn't get it to work until I changed the clk to 100 MHz.

I could mark nets for debug, complete the 'Set Up debug' wizard and complete bitstream generation but nothing would happen in the waveform view.

I tried various IP example projects and found I could debug them without a problem.

After playing with the ILA IP example I now believe there is a minimum clk frequency for the debug IP.

From Vivado Design Suite User Guide Programming and Debugging UG908:
    IMPORTANT: Ensure the JTAG clock is slower than the clocks input to the debug cores.

When I first read this, I got it the wrong way around and thought the design clock needed to be less than 15 MHz. I now realise it needs to be greater than 15 MHz.

By trial and error with the ILA example it appears the minimum design clk is 20 MHz.

The JTAG frequency in Vivado defaults to 15 MHz. 'FREQUENCY' is one of the properties of the hardware target, which can be seen in the Hardware Manager.

The frequency of the LED flash has changed a bit but it was always pretty arbitrary.